[#5] update build gradle
