+++ spring-web-kotlin-blocking/src/test/kotlin/kim/myeongjae/spring_web_kotiln_blocking/article/api/ArticleControllerTest.kt
@@ -3,10 +3,8 @@ package kim.myeongjae.spring_web_kotiln_blocking.article.api
 import kim.myeongjae.spring_web_kotiln_blocking.article.domain.model.ArticleFixture
 import kim.myeongjae.spring_web_kotiln_blocking.article.domain.model.ArticleRepository
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.ArgumentMatchers
 import org.mockito.BDDMockito
-import org.mockito.junit.jupiter.MockitoExtension
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
@@ -22,7 +20,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 
 @WebMvcTest(ArticleController::class)
 @AutoConfigureRestDocs
-@ExtendWith(MockitoExtension::class)
 class ArticleControllerTest @Autowired constructor(
     private val mvc: MockMvc,
 ) {
@@ -40,18 +37,19 @@ class ArticleControllerTest @Autowired constructor(
         mvc.perform(RestDocumentationRequestBuilders.get("/articles/{slug}", slug))
             .andExpect(MockMvcResultMatchers.status().isOk)
             .andDo(MockMvcResultHandlers.print())
-            .andDo(MockMvcRestDocumentation.document(
-                "articles/get-external",
-                Preprocessors.preprocessRequest(Preprocessors.prettyPrint()),
-                Preprocessors.preprocessResponse(Preprocessors.prettyPrint()),
-                RequestDocumentation.pathParameters(RequestDocumentation.parameterWithName("slug").description("슬러그")),
-                PayloadDocumentation.responseFields(
-                    PayloadDocumentation.fieldWithPath("title").description("제목"),
-                    PayloadDocumentation.fieldWithPath("content").description("내용"),
-                    PayloadDocumentation.fieldWithPath("createdAt").description("생성일시"),
-                    PayloadDocumentation.fieldWithPath("updatedAt").description("갱신일시"),
-                ),
-            ))
-
+            .andDo(
+                MockMvcRestDocumentation.document(
+                    "articles/get-external",
+                    Preprocessors.preprocessRequest(Preprocessors.prettyPrint()),
+                    Preprocessors.preprocessResponse(Preprocessors.prettyPrint()),
+                    RequestDocumentation.pathParameters(RequestDocumentation.parameterWithName("slug").description("슬러그")),
+                    PayloadDocumentation.responseFields(
+                        PayloadDocumentation.fieldWithPath("title").description("제목"),
+                        PayloadDocumentation.fieldWithPath("content").description("내용"),
+                        PayloadDocumentation.fieldWithPath("createdAt").description("생성일시"),
+                        PayloadDocumentation.fieldWithPath("updatedAt").description("갱신일시"),
+                    ),
+                )
+            )
     }
 }
